best/requirement.txt
torch
python-chess
zstandard
tqdm
numpy